requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
streamlit>=1.32.0
langchain>=0.1.13
langchain-openai>=0.1.0
langchain-groq
langchain-community>=0.0.29
chromadb>=0.4.24
sentence-transformers>=2.6.1
pydantic>=2.6.4
python-dotenv>=1.0.1
PyPDF2>=3.0.1
